requirements.txt
fastapi[all]
pydantic
manim==0.18.0
ffmpeg-python
redis
google-generativeai
python-dotenv
numpy